middleman/requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
black==23.10.1
blinker==1.6.3
blis==0.7.11
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
chroma-hnswlib==0.7.3
chromadb==0.4.17
click==8.1.7
cloudpathlib==0.16.0
coloredlogs==15.0.1
confection==0.1.3
cryptography==41.0.5
cymem==2.0.8
dataclasses-json==0.6.1
Deprecated==1.2.14
dill==0.3.7
dnspython==2.4.2
elastic-transport==8.4.1
elasticsearch==8.10.1
faiss-cpu==1.7.4
fastapi==0.104.1
filelock==3.12.4
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.10.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.12.0
google-api-python-client==2.105.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.0
grpcio-tools==1.59.0
h11==0.14.0
h2==4.1.0
h5py==3.10.0
hpack==4.0.0
httpcore==0.18.0
httplib2==0.22.0
httptools==0.6.1
httpx==0.25.0
huggingface-hub==0.17.3
humanfriendly==10.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
InstructorEmbedding==1.0.1
jaraco.context==4.3.0
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
keras==2.14.0
kubernetes==28.1.0
langchain==0.0.335
langcodes==3.3.0
langsmith==0.0.64
libclang==16.0.6
loguru==0.7.2
manifest-ml==0.1.8
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
ml-dtypes==0.2.0
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
mypy==1.6.1
mypy-extensions==1.0.0
networkx==3.2
nltk==3.8.1
numpy==1.26.1
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.9.4.25
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.16.2
openai==0.28.1
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.2
pandas==2.1.2
pathspec==0.11.2
Pillow==10.1.0
pinecone-client==2.2.4
platformdirs==3.11.0
portalocker==2.8.2
posthog==3.0.2
preshed==3.0.9
protobuf==4.24.4
pulsar-client==3.3.0
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
PyPDF2==3.0.1
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
qdrant-client==1.6.3
redis==5.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.22
sqlitedict==2.1.0
srsly==2.4.8
starlette==0.27.0
streamlit==1.28.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-hub==0.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-text==2.14.0
termcolor==2.3.0
thinc==8.2.1
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.14.1
toml==0.10.2
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
transformers==4.34.1
triton==2.1.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.23.2
uvloop==0.19.0
validators==0.22.0
wasabi==1.1.2
watchdog==3.0.0
watchfiles==0.21.0
weasel==0.3.3
weaviate-client==3.24.2
websocket-client==1.6.4
websockets==12.0
Werkzeug==3.0.1
wikipedia==1.4.0
wolframalpha==5.0.0
wrapt==1.14.1
xmltodict==0.13.0
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
